pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vitalis"
version = "0.1.0"
description = "A CLI tool for extracting and analyzing dependency manifests across multiple ecosystems"
readme = "README.md"
license = "Apache-2.0"
authors = [
    {name = "Vitalis Team"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
requires-python = ">=3.9"
dependencies = [
    "typer>=0.9.0",
    "pyyaml>=6.0",
    "toml>=0.10.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "build>=0.10.0",
    "twine>=4.0.0",
    "jsonschema>=4.0.0",
    "mypy>=1.0.0",
    "coverage>=7.0.0",
    "types-toml>=0.10.0",
    "types-PyYAML>=6.0.0",
]

[project.scripts]
vitalis = "extractor.cli:app"

[project.urls]
"Homepage" = "https://github.com/zchryr/vitalis"
"Bug Reports" = "https://github.com/zchryr/vitalis/issues"
"Source" = "https://github.com/zchryr/vitalis"

[tool.setuptools.packages.find]
include = ["extractor*", "analyzer*", "core*"]
exclude = ["tests*"]